다이어리 상세조회/삭제/수정 완료

+++ lib/feature/diary/presentation/viewmodels/edit_diary_state.dart
@@ -75,7 +75,7 @@ abstract class EditDiaryState with _$EditDiaryState {
       type: type,
       title: title.isNotEmpty ? title : null,
       content: content.isNotEmpty ? content : null,
-      rating: type == 'REVIEW' ? rating.toInt() : null,
+      rating: type == 'REVIEW' ? rating.toDouble() : null,
       img: type == 'PHOTO' ? imgUrl : null,
       cost: type == 'MONEY' ? cost : null,
       isPublic: isPublic,
